requirements.txt
# Core dependencies
playwright>=1.40.0
pyyaml>=6.0.1

# API Server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# HTTP client for webhooks
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0

# Optional: Airtable integration (future use)
pyairtable>=2.1.0

# Optional: PostgreSQL integration (future use)
psycopg2-binary>=2.9.9
